rip it all apart and refactor
