+++ distributaur/__init__.py
@@ -1,3 +1,1 @@
-from .core import *
-from .vast import *
-from .huggingface import *
+from .distributaur import *
