+++ distributaur/tests/__init__.py
@@ -1,3 +1,1 @@
-from .core_test import *
-from .vast_test import *
-from .huggingface_test import *
+from .distributaur_test import *
