+++ distributaur/tests/test_worker.py
@@ -1,4 +1,6 @@
-from distributaur.core import register_function, app
+from distributaur.distributaur import Distributaur
+
+distributaur = Distributaur()
 
 
 # Define and register the example_function
@@ -6,4 +8,4 @@ def example_function(arg1, arg2):
     return f"Result: arg1={arg1}, arg2={arg2}"
 
 
-register_function(example_function)
+distributaur.register_function(example_function)
