+++ example_worker.py
@@ -3,7 +3,10 @@ import sys
 
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "./"))
 
-from distributaur.core import app
-import example
+from distributaur.distributaur import Distributaur
 
-celery = app
+from example import *
+
+if __name__ == "__main__":
+    distributaur = Distributaur()
+    celery = distributaur.app
